Split UtilTest into Array and Http. Fixes #2 on Out of Bounds exception
